Go through effect differences for Aus

+++ src/effect/ef_gimonhu.c
@@ -92,7 +92,7 @@ static void eGM_dw(eEC_Effect_c* effect, GAME* game) {
     _texture_z_light_fog_prim_xlu(game->graph);
     eEC_CLIP->auto_matrix_xlu_offset_proc(game, &effect->position, &effect->scale, &effect->offset);
     gSPMatrix(NEXT_POLY_XLU_DISP, _Matrix_to_Mtx_new(game->graph), G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_MODELVIEW);
-    gDPSetPrimColor(NEXT_POLY_XLU_DISP, 0, 128, 255, 255, 255, v);
+    gDPSetPrimColor(NEXT_POLY_XLU_DISP, 0, 128, 255, 255, 255, (u8)v);
     gSPDisplayList(NEXT_POLY_XLU_DISP, ef_gimonhu01_00_modelT);
     CLOSE_DISP(game->graph);
 }
