+++ src/effect/ef_goki.c
@@ -83,7 +83,7 @@ static void eGoki_dw(eEC_Effect_c* effect, GAME* game) {
     Matrix_RotateZ(effect->effect_specific[0], MTX_MULT);
     Evw_Anime_Set(play, ef_goki01_01_evw_anime);
     gSPMatrix(NEXT_POLY_XLU_DISP, _Matrix_to_Mtx_new(game->graph), G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_MODELVIEW);
-    gDPSetPrimColor(NEXT_POLY_XLU_DISP, 0, 255, 255, 255, 255, opacity);
+    gDPSetPrimColor(NEXT_POLY_XLU_DISP, 0, 255, 255, 255, 255, (u8)opacity);
     gSPDisplayList(NEXT_POLY_XLU_DISP, ef_goki01_01_modelT);
     CLOSE_DISP(game->graph);
 }
