+++ src/effect/ef_ha.c
@@ -69,7 +69,7 @@ static void eHA_dw(eEC_Effect_c* effect, GAME* game) {
                  effect->scale.y * (GETREG(MYKREG, 0x1b) * 0.01f + 1.f),
                  effect->scale.z * (GETREG(MYKREG, 0x1b) * 0.01f + 1.f), MTX_MULT);
     gSPMatrix(NEXT_POLY_XLU_DISP, _Matrix_to_Mtx_new(game->graph), G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_MODELVIEW);
-    gDPSetPrimColor(NEXT_POLY_XLU_DISP, 0, 128, 255, 255, 255, opacity);
+    gDPSetPrimColor(NEXT_POLY_XLU_DISP, 0, 128, 255, 255, 255, (u8)opacity);
     gDPSetEnvColor(NEXT_POLY_XLU_DISP, 0, 0, 255, 255);
     gSPDisplayList(NEXT_POLY_XLU_DISP, ef_ha01_00_modelT);
     CLOSE_DISP(game->graph);
