+++ src/effect/ef_neboke_awa.c
@@ -105,7 +105,7 @@ static void eSleep_dw(eEC_Effect_c* effect, GAME* game) {
     Matrix_scale(effect->scale.x, effect->scale.y, effect->scale.z, MTX_MULT);
 
     gSPMatrix(NEXT_POLY_XLU_DISP, _Matrix_to_Mtx_new(game->graph), G_MTX_NOPUSH | G_MTX_LOAD | G_MTX_MODELVIEW);
-    gDPSetPrimColor(NEXT_POLY_XLU_DISP, 0, 128, 255, 255, 255, effect->effect_specific[1]);
+    gDPSetPrimColor(NEXT_POLY_XLU_DISP, 0, 128, 255, 255, 255, (u8)effect->effect_specific[1]);
     gSPDisplayList(NEXT_POLY_XLU_DISP, ef_sleep01_modelT);
 
     effect->offset = vec;
